pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "coding-assistants"
version = "0.1.0"
description = "An environment where multiple LLMs can talk among each other to help program a given codebase"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "openai>=1.0.0",
    "pyautogen>=0.2.0",
    "python-dotenv>=1.0.0",
    "rich>=13.0.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]

# --- Pytest Configuration ---
[tool.pytest.ini_options]
# Specify the directory where tests are located
testpaths = [
    "tests",
]
# Specify patterns for test files and functions
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"